fix(Modal.stories): Modal, activeModal 스토리를 하나로 병합
